DataAnalysisSystem/requirements.txt
click==8.1.3
colorama==0.4.4
cycler==0.11.0
Flask==2.1.2
fonttools==4.43.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==11.1.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
Werkzeug==2.1.2
wordcloud==1.9.4
zipp==3.19.1
SQLAlchemy==1.4.3

